backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
langchain
langchain-core
langchain-community
langchain-anthropic
anthropic
openai
pydantic
email-validator
websockets
twilio
python-multipart
aiofiles
beautifulsoup4
playwright
arq
redis
numpy
